Holo.Light --> DeviceDefault
I remember running the ROM a while ago, and there was one case where the text was white-on-white, making it difficult to read.

I know you can't change a ListViews textColor unless you switch the whole theme of the app, so I'm thinking this may be a good fix. Holo Light looks elegant, dont get me wrong, but it almost seems 'out of place' in the Settings app because the background of this ROMs settings is the stock color. So, using DeviceDefault might work better.